+++ .gitignore
@@ -4,9 +4,13 @@
 *.dll
 *.so
 *.dylib
+.DS_Store
 
 # Test binary, build with `go test -c`
 *.test
 
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
+
+vendor/
+openrc
