+++ README.md
@@ -1,28 +1,35 @@
 # OpenStack-Exporter
+Simple Prometheus exporter for OpenStack.
 
-Simple Prometheus exporter for OpenStack
+## Build
+First clone the repo, and excute `Makefile` using make tool:
+```sh
+$ git clone https://github.com/iGene/openstack-exporter.git $GOPATH/src/github.com/iGene/openstack-exporter
+$ cd $GOPATH/src/github.com/iGene/openstack-exporter
+$ make
+```
+
+A Dockerfile is provided in this repo, to build a OpenStack Exporter Docker image, simply run:
+```sh
+$ make build_image
+```
 
 ## Deployment
-
-First clone the repo
-
-```git clone https://github.com/iGene/openstack-exporter.git```
-
-A Dockerfile is provided in this repo, to build a OpenStack Exporter Docker image, simply run
-
-```docker build -t openstack-exporter openstack-exporter/```
-
-Copy the sample configuraion file and fill in
+Copy the sample configuraion file and fill in:
 
 - OpenStack Username
 - OpenStack Password
 - OpenStack Keystone Endpoint
-- OpenStack Project Name
 
-Launch the exporter using Docker
+Launch the exporter using Docker:
+```sh
+$ docker run -d -p 9183:9183  \
+      -v $(pwd)/openstack.toml:/etc/openstack-exporter/openstack.toml \
+      --name openstack-exporter \
+      openstack-exporter:v0.1.0
+```
 
-```docker run --name openstack-exporter -v $(pwd)/openstack.toml:/etc/openstack-exporter/openstack.toml -d -p 9183:9183 openstack-exporter```
-
-Check if its working by
-
-```curl localhost:9183/metrics```
+Check if its working by:
+```sh
+$ curl localhost:9183/metrics
+```
