+++ deploy/exporter-sm.yml
@@ -1,0 +1,17 @@
+apiVersion: monitoring.coreos.com/v1
+kind: ServiceMonitor
+metadata:
+  name: openstack-exporter
+  labels:
+    k8s-app: openstack-exporter
+  namespace: monitoring
+spec:
+  selector:
+    matchLabels:
+      k8s-app: openstack-exporter
+  namespaceSelector:
+    matchNames:
+    - monitoring
+  endpoints:
+  - port: http-metrics
+    interval: 30s
